factor out common hooks
